refactor: redistribute Issue*Test files into thematic OpenCypher test classes
Move 22 Issue*Test regression tests from standalone files into their
corresponding feature-area test classes as @Nested inner classes, with
issue references preserved in javadoc. Each nested class retains its own
DB lifecycle to avoid interference with the outer class.

Mapping:
- OpenCypherWhereClauseTest ← #3128, #3132, #3320, #3336, #3403
- OpenCypherCollectUnwindTest ← #3129, #3138, #3154, #3612
- CollectDistinctTest ← #3316, #3404, #3271
- OpenCypherMergeTest ← #3131
- OpenCypherOptionalMatchTest ← #3218
- OpenCypherPatternPredicateTest ← #3331
- OpenCypherStringEscapeTest ← #3333
- CypherLabelFilteringTest ← #3412
- OpenCypherBasicTest ← #3415
- OpenCypherTraversalTest ← #3285
- OpenCypherUnionCallProfileTest ← #3407
- OpenCypherMathFunctionsTest ← #3417
- NEW OpenCypherSpatialFunctionsTest ← #3402

Issue3216Test and Issue3359Test kept standalone (performance/concurrency).
